requirements.txt
Django==4.0.3
python-dotenv==0.20.0
psycopg2==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
praw==7.5.0
djangorestframework-simplejwt==5.2.0
cachetools==5.2.0
whitenoise==6.2.0
gunicorn==20.1.0
